chore: bump version to 0.9.4; build js for docs with bun as well
