+++ docs/assets/js/main.js
@@ -1,636 +1,1 @@
-import {
-	UIElement, maybe, pass, on, effect,
-	asBoolean, asInteger,
-	setText, setProperty, setAttribute, toggleAttribute, toggleClass
-} from '@efflore/ui-element'
-import Prism from './prism.min.js'
-
-class MyCounter extends UIElement {
-	static observedAttributes = ['count']
-	static attributeMap = {
-        count: asInteger
-    }
-
-	connectedCallback() {
-		this.set('parity', () => this.get('count') % 2 ? 'odd' : 'even')
-        this.first('.increment').forEach(on('click', () => this.set('count', v => ++v)))
-        this.first('.decrement').forEach(on('click', () => this.set('count', v => --v)))
-		this.first('.count').forEach(setText('count'))
-		this.first('.parity').forEach(setText('parity'))
-    }
-}
-MyCounter.define('my-counter')
-
-class HelloWorld extends UIElement {
-	connectedCallback() {
-        this.first('span').forEach(setText('name'))
-		this.first('input').forEach(on('input', e => this.set('name', e.target.value || undefined)))
-	}
-}
-HelloWorld.define('hello-world')
-
-class MySlider extends UIElement {
-	connectedCallback() {
-	
-		// Initialize state for the active slide index
-		this.set('active', 0)
-	
-		// Event listeners for navigation
-		const total = this.querySelectorAll('.slide').length
-		const getNewIndex = (prev, direction) => (prev + direction + total) % total
-		this.first('.prev').map(on('click', () => this.set('active', v => getNewIndex(v, -1))))
-		this.first('.next').map(on('click', () => this.set('active', v => getNewIndex(v, 1))))
-	
-		// Auto-effects for updating slides and dots
-		this.all('.slide').map((ui, idx) => toggleClass('active', () => idx === this.get('active'))(ui))
-		this.all('.dots span').map((ui, idx) => toggleClass('active', () => idx === this.get('active'))(ui))
-	}
-}
-MySlider.define('my-slider')
-
-class CodeBlock extends UIElement {
-	static observedAttributes = ['collapsed']
-	static attributeMap = {
-		collapsed: asBoolean
-	}
-
-  	connectedCallback() {
-
-		// Enhance code block with Prism.js
-		const language = this.getAttribute('language') || 'html'
-		const content = this.querySelector('code')
-		this.set('code', content.textContent.trim(), false)
-		effect(enqueue => {
-			// Apply syntax highlighting while preserving Lit's marker nodes in Storybook
-			const code = document.createElement('code')
-			code.innerHTML = Prism.highlight(this.get('code'), Prism.languages[language], language)
-			enqueue(content, 'h', el => () => {
-				Array.from(el.childNodes)
-					.filter(node => node.nodeType !== Node.COMMENT_NODE)
-					.forEach(node => node.remove())
-				Array.from(code.childNodes)
-					.forEach(node => el.appendChild(node))
-			})
-		})
-
-		// Copy to clipboard
-		this.first('.copy').map(ui => on('click', async () => {
-			const copyButton = ui.target
-			const label = copyButton.textContent
-			let status = 'success'
-			try {
-				await navigator.clipboard.writeText(content.textContent)
-			} catch (err) {
-				console.error('Error when trying to use navigator.clipboard.writeText()', err)
-				status = 'error'
-			}
-			copyButton.set('disabled', true)
-			copyButton.set('label', ui.host.getAttribute(`copy-${status}`))
-			setTimeout(() => {
-				copyButton.set('disabled', false)
-				copyButton.set('label', label)
-			}, status === 'success' ? 1000 : 3000)
-		})(ui))
-
-		// Expand
-		this.first('.overlay').map(on('click', () => this.set('collapsed', false)))
-		this.self.map(toggleAttribute('collapsed'))
-	}
-}
-CodeBlock.define('code-block')
-
-class TabList extends UIElement {
-	static observedAttributes = ['accordion']
-	static attributeMap = {
-        accordion: asBoolean
-    }
-	static consumedContexts = ['media-viewport']
-
-	connectedCallback() {
-		super.connectedCallback()
-		this.set('active', 0, false) // initial active tab
-
-		// Dynamically adjust accordion based on viewport size
-		setTimeout(() => {
-			if (this.get('media-viewport'))
-				this.set('accordion', () => ['xs', 'sm'].includes(this.get('media-viewport')))
-		}, 0)
-
-		// Reflect accordion attribute (may be used for styling)
-		this.self.forEach(toggleAttribute('accordion'))
-
-		// Hide accordion tab navigation when in accordion mode
-		this.first('menu').forEach(setProperty('ariaHidden', 'accordion'))
-
-		// Update active tab state and bind click handlers
-		this.all('menu button')
-			.map((ui, idx) => setProperty('ariaPressed', () => this.get('active') === idx)(ui))
-			.forEach((ui, idx) => on('click', () => this.set('active', idx))(ui))
-
-		// Pass open and collapsible states to accordion panels
-		this.all('accordion-panel').forEach((ui, idx) => pass({
-			open: () => ui.host.get('active') === idx,
-			collapsible: 'accordion'
-		})(ui))
-	}
-}
-TabList.define('tab-list')
-
-class AccordionPanel extends UIElement {
-	connectedCallback() {
-
-		// Set defaults from attributes
-		this.set('open', this.hasAttribute('open'), false)
-		this.set('collapsible', this.hasAttribute('collapsible'), false)
-
-		// Handle open and collapsible state changes
-		this.self
-		    .map(toggleAttribute('open'))
-			.map(toggleAttribute('collapsible'))
-			.forEach(setProperty('ariaHidden', () => !this.get('open') && !this.get('collapsible')))
-
-		// Control inner details panel
-		this.first('details')
-			.map(setProperty('open'))
-			.forEach(setProperty('ariaDisabled', () => !this.get('collapsible')))
-	}
-}
-AccordionPanel.define('accordion-panel')
-
-class InputButton extends UIElement {
-	static observedAttributes = ['disabled']
-	static attributeMap = {
-		disabled: asBoolean
-	}
-
-	connectedCallback() {
-		this.first('button')
-			.map(setText('label'))
-			.map(setProperty('disabled'))
-	}
-}
-InputButton.define('input-button')
-
-const isNumber = num => typeof num === 'number'
-const parseNumber = (v, int = false) => int ? parseInt(v, 10) : parseFloat(v)
-
-class InputField extends UIElement {
-	static observedAttributes = ['value', 'description']
-	static attributeMap = {
-		value: value => this.isNumber
-			? value.map(v => parseNumber(v, this.isInteger)).filter(Number.isFinite)
-			: value
-	}
-
-	connectedCallback() {
-		this.input = this.querySelector('input')
-		this.isNumber = this.input && this.input.type === 'number'
-		this.isInteger = this.hasAttribute('integer')
-
-		// Set default states
-		this.set('value', this.isNumber ? this.input.valueAsNumber : this.input.value, false)
-		this.set('length', this.input.value.length)
-		
-		// Derived states
-		this.set('empty', () => !this.get('length'))
-
-		// Setup sub elements
-		this.#setupErrorMessage()
-		this.#setupDescription()
-		this.#setupSpinButton()
-		this.#setupClearButton()
-
-		// Handle input changes
-		this.input.onchange = () => this.#triggerChange(this.isNumber ? this.input.valueAsNumber : this.input.value)
-		this.input.oninput = () => this.set('length', this.input.value.length)
-
-		// Update value
-		effect(async () => {
-			const value = this.get('value')
-			const validate = this.getAttribute('validate')
-			if (value && validate) {
-
-				// Validate input value against a server-side endpoint
-				await fetch(`${validate}?name=${this.input.name}value=${this.input.value}`)
-				.then(async response => {
-					const text = await response.text()
-					this.input.setCustomValidity(text)
-					this.set('error', text)
-				})
-				.catch(err => this.set('error', err.message))
-			}
-			if (this.isNumber && !isNumber(value)) // ensure value is a number if it is not already a number
-				return this.set('value', parseNumber(value, this.isInteger)) // effect will be called again with numeric value
-			if (this.isNumber && !Number.isNaN(value)) // change value only if it is a valid number
-				this.input.value = value
-		})
- 	}
-
-	// Clear the input field
-	clear() {
-		this.input.value = ''
-		this.set('value', '')
-		this.set('length', 0)
-		this.input.focus()
-	}
-
-	// Trigger value-change event to commit the value change
-	#triggerChange = value => {
-		this.set('value', value)
-		this.set('error', this.input.validationMessage)
-		if (typeof value === 'function')
-			value = this.get('value')
-		if (this.input.value !== String(value))
-			this.dispatchEvent(new CustomEvent('value-change', {
-				detail: value,
-				bubbles: true
-			}))
-	}
-
-	// Setup error message
-	#setupErrorMessage() {
-		const error = this.first('.error')
-
-		// Derived states
-		this.set('ariaInvalid', () => String(Boolean(this.get('error'))))
-		this.set('aria-errormessage', () => this.get('error') ? error[0]?.target.id : undefined)
-
-		// Effects
-		error.forEach(setText('error'))
-		this.first('input')
-			.map(setProperty('ariaInvalid'))
-			.forEach(setAttribute('aria-errormessage'))
-	}
-
-	// Setup description
-	#setupDescription() {
-		const description = this.first('.description')
-		if (!description[0])
-			return // no description, so skip
-		
-		// derived states
-		const input = this.first('input')
-		const maxLength = this.input.maxLength
-		const remainingMessage = maxLength && description[0].target.dataset.remaining
-		const defaultDescription = description[0].target.textContent
-		this.set('description', remainingMessage
-			? () => {
-				const length = this.get('length')
-				return length > 0
-					? remainingMessage.replace('${x}', maxLength - length)
-					: defaultDescription
-			}
-			: defaultDescription
-		)
-		this.set('aria-describedby', () => this.get('description')
-			? description[0].target.id
-			: undefined
-		)
-
-		// Effects
-		description.forEach(setText('description'))
-		input.forEach(setAttribute('aria-describedby'))
-	}
-
-	// Setup spin button
-	#setupSpinButton() {
-		const spinButton = this.querySelector('.spinbutton')
-		if (!this.isNumber || !spinButton)
-			return // no spin button, so skip
-
-		const getNumber = attr => maybe(parseNumber(this.input[attr], this.isInteger)).filter(Number.isFinite)[0]
-		const tempStep = parseNumber(spinButton.dataset.step, this.isInteger)
-		const [step, min, max] = Number.isFinite(tempStep)
-			? [tempStep, getNumber('min'), getNumber('max')]
-			: []
-
-		// Bring value to nearest step
-		const nearestStep = v => {
-			const steps = Math.round((max - min) / step)
-			let zerone = Math.round((v - min) * steps / (max - min)) / steps // bring to 0-1 range
-			zerone = Math.min(Math.max(zerone, 0), 1) // keep in range in case value is off limits
-			const value = zerone * (max - min) + min
-			return this.isInteger ? Math.round(value) : value
-		}
-
-		// Step down
-		this.stepDown = (stepDecrement = step) => this.#triggerChange(v => nearestStep(v - stepDecrement))
-
-		// Step up
-		this.stepUp = (stepIncrement = step) => this.#triggerChange(v => nearestStep(v + stepIncrement))
-
-		// derived states
-		this.set('decrement-disabled', () => isNumber(min) && (this.get('value') - step < min))
-		this.set('increment-disabled', () => isNumber(max) && (this.get('value') + step > max))
-
-		// Handle spin button clicks and update their disabled state
-		this.first('.decrement')
-			.map(setProperty('disabled', 'decrement-disabled'))
-			.forEach(on('click', e => this.stepDown(e.shiftKey ? step * 10 : step)))
-		this.first('.increment')
-			.map(setProperty('disabled', 'increment-disabled'))
-			.forEach(on('click', e => this.stepUp(e.shiftKey ? step * 10 : step)))
-
-		// Handle arrow key events
-		this.input.onkeydown = e => {
-			if (['ArrowUp', 'ArrowDown'].includes(e.key)) {
-				e.stopPropagation()
-				e.preventDefault()
-				if (e.key === 'ArrowDown')
-					this.stepDown(e.shiftKey ? step * 10 : step)
-				if (e.key === 'ArrowUp')
-					this.stepUp(e.shiftKey ? step * 10 : step)
-			}
-		}
-	}
-
-	// Setup clear button
-	#setupClearButton() {
-		this.first('.clear')
-			.map(toggleClass('hidden', 'empty'))
-			.forEach(on('click', () => {
-				this.clear()
-				this.input.focus()
-			}))
-	}
-
-}
-InputField.define('input-field')
-
-class InputCheckbox extends UIElement {
-	static observedAttributes = ['checked']
-	static attributeMap = {
-		checked: asBoolean,
-	}
-
-	connectedCallback() {
-		this.first('input')
-			.map(on('change', e => this.set('checked', Boolean(e.target.checked))))
-			.forEach(setProperty('checked'))
-		this.self.forEach(toggleAttribute('checked'))
-	}
-}
-InputCheckbox.define('input-checkbox')
-
-export class InputRadiogroup extends UIElement {
-	static observedAttributes = ['value']
-
-	connectedCallback() {
-		this.self.forEach(setAttribute('value'))
-        this.all('input').forEach(on('change', e => this.set('value', e.target.value)))
-		this.all('label').forEach(ui => toggleClass(
-			'selected',
-			() => ui.host.get('value') === ui.target.querySelector('input').value
-		)(ui))
-    }
-}
-InputRadiogroup.define('input-radiogroup')
-
-class LazyLoad extends UIElement {
-	static observedAttributes = ['src']
-	static attributeMap = {
-		src: v => v.map(src => {
-				let url = ''
-				try {
-					url = new URL(src, location.href) // ensure 'src' attribute is a valid URL
-					if (url.origin !== location.origin) { // sanity check for cross-origin URLs
-						throw new TypeError('Invalid URL origin')
-					}
-				} catch (error) {
-					console.error(error, url)
-					url = ''
-				}
-				return url.toString()
-			})
-	}
-
-	connectedCallback() {
-
-		// Show / hide loading message
-		this.first('.loading')
-			.forEach(setProperty('ariaHidden', () => !!this.get('error')))
-
-		// Set and show / hide error message
-		this.first('.error')
-			.map(setText('error'))
-			.forEach(setProperty('ariaHidden', () => !this.get('error')))
-
-		// Load content from provided URL
-		effect(enqueue => {
-			const src = this.get('src')
-			if (!src) return // silently fail if no valid URL is provided
-			fetch(src)
-				.then(async response => {
-					if (response.ok) {
-						const content = await response.text()
-						enqueue(this.root, 'h', el => () => {
-							// UNSAFE!, use only trusted sources in 'src' attribute
-							el.innerHTML = content
-							el.querySelectorAll('script').forEach(script => {
-								const newScript = document.createElement('script')
-								newScript.appendChild(document.createTextNode(script.textContent))
-								el.appendChild(newScript)
-								script.remove()
-							})
-						})
-						this.set('error', '')
-					} else {
-						this.set('error', response.status + ':'+ response.statusText)
-					}
-				})
-				.catch(error => this.set('error', error))
-		})
-	}
-}
-LazyLoad.define('lazy-load')
-
-const MEDIA_MOTION = 'media-motion'
-const MEDIA_THEME = 'media-theme'
-const MEDIA_VIEWPORT = 'media-viewport'
-const MEDIA_ORIENTATION = 'media-orientation'
-
-class MediaContext extends UIElement {
-	static providedContexts = [MEDIA_MOTION, MEDIA_THEME, MEDIA_VIEWPORT, MEDIA_ORIENTATION]
-
-	connectedCallback() {
-		super.connectedCallback()
-
-		const THEME_LIGHT = 'light'
-		const THEME_DARK = 'dark'
-		const VIEWPORT_XS = 'xs'
-		const VIEWPORT_SM = 'sm'
-		const VIEWPORT_MD = 'md'
-		const VIEWPORT_LG = 'lg'
-		const VIEWPORT_XL = 'xl'
-		const ORIENTATION_LANDSCAPE = 'landscape'
-		const ORIENTATION_PORTRAIT = 'portrait'
-
-		const getBreakpoints = () => {
-			const parseBreakpoint = breakpoint => {
-				const attr = this.getAttribute(breakpoint)?.trim()
-				if (!attr) return null
-				const unit = attr.match(/em$/) ? 'em' : 'px'
-				const value = maybe(parseFloat(attr)).filter(Number.isFinite)[0]
-				return value ? value + unit : null
-			}
-			const sm = parseBreakpoint(VIEWPORT_SM) || '32em'
-			const md = parseBreakpoint(VIEWPORT_MD) || '48em'
-			const lg = parseBreakpoint(VIEWPORT_LG) || '72em'
-			const xl = parseBreakpoint(VIEWPORT_XL) || '108em'
-			return { sm, md, lg, xl }
-		}
-		const breakpoints = getBreakpoints()
-
-		const reducedMotion = matchMedia('(prefers-reduced-motion: reduce)')
-		const colorScheme = matchMedia('(prefers-color-scheme: dark)')
-		const screenSmall = matchMedia(`(min-width: ${breakpoints.sm})`)
-		const screenMedium = matchMedia(`(min-width: ${breakpoints.md})`)
-		const screenLarge = matchMedia(`(min-width: ${breakpoints.lg})`)
-		const screenXLarge = matchMedia(`(min-width: ${breakpoints.xl})`)
-		const screenOrientation = matchMedia('(orientation: landscape)')
-
-		const getViewport = () => {
-			if (screenXLarge.matches) return VIEWPORT_XL
-            if (screenLarge.matches) return VIEWPORT_LG
-            if (screenMedium.matches) return VIEWPORT_MD
-            if (screenSmall.matches) return VIEWPORT_SM
-            return VIEWPORT_XS
-		}
-
-		// set initial values
-		this.set(MEDIA_MOTION, reducedMotion.matches)
-		this.set(MEDIA_THEME, colorScheme.matches ? THEME_DARK : THEME_LIGHT)
-		this.set(MEDIA_VIEWPORT, getViewport())
-		this.set(MEDIA_ORIENTATION, screenOrientation.matches ? ORIENTATION_LANDSCAPE : ORIENTATION_PORTRAIT)
-
-		// event listeners
-		reducedMotion.addEventListener(
-			'change',
-			e => this.set(MEDIA_MOTION, e.matches)
-		)
-		colorScheme.addEventListener(
-			'change',
-			e => this.set(MEDIA_THEME, e.matches ? THEME_DARK : THEME_LIGHT)
-		)
-		screenSmall.addEventListener('change', () => this.set(MEDIA_VIEWPORT, getViewport()))
-		screenMedium.addEventListener('change', () => this.set(MEDIA_VIEWPORT, getViewport()))
-		screenLarge.addEventListener('change', () => this.set(MEDIA_VIEWPORT, getViewport()))
-		screenXLarge.addEventListener('change', () => this.set(MEDIA_VIEWPORT, getViewport()))
-		screenOrientation.addEventListener(
-			'change',
-			e => this.set(MEDIA_THEME, e.matches ? ORIENTATION_LANDSCAPE : ORIENTATION_PORTRAIT)
-		)
-	}
-}
-MediaContext.define('media-context')
-
-class TodoApp extends UIElement {
-	connectedCallback() {
-		const [todoList, todoFilter] = ['todo-list', 'input-radiogroup']
-			.map(selector => this.querySelector(selector))
-
-		// Event listener on own element
-		this.self.forEach(on('add-todo', e => todoList?.addItem(e.detail)))
-		
-		// Coordinate todo-count
-		this.first('todo-count').forEach(pass({
-			active: () => todoList?.get('count').active
-		}))
-
-		// Coordinate todo-list
-		this.first('todo-list').forEach(pass({
-			filter: () => todoFilter?.get('value')
-		}))
-
-		// Coordinate .clear-completed button
-		this.first('.clear-completed')
-			.map(on('click', () => todoList?.clearCompleted()))
-			.forEach(pass({ disabled: () => !todoList?.get('count').completed }))
-	}
-}
-TodoApp.define('todo-app')
-
-class TodoCount extends UIElement {
-	connectedCallback() {
-		this.set('active', 0, false)
-		this.first('.count').forEach(setText('active'))
-		this.first('.singular').forEach(setProperty('ariaHidden', () => this.get('active') > 1))
-		this.first('.plural').forEach(setProperty('ariaHidden', () => this.get('active') === 1))
-		this.first('.remaining').forEach(setProperty('ariaHidden', () => !this.get('active')))
-		this.first('.all-done').forEach(setProperty('ariaHidden', () => !!this.get('active')))
-	}
-}
-TodoCount.define('todo-count')
-
-class TodoForm extends UIElement {
-	connectedCallback() {
-		const inputField = this.querySelector('input-field')
-
-        this.first('form').forEach(on('submit', e => {
-			e.preventDefault()
-			setTimeout(() => {
-				this.dispatchEvent(new CustomEvent('add-todo', {
-					bubbles: true,
-					detail: inputField.get('value')
-				}))
-				inputField.clear()
-			}, 0)
-		}))
-	
-		this.first('input-button').forEach(pass({
-			disabled: () => inputField.get('empty')
-		}))
-    }
-}
-TodoForm.define('todo-form')
-
-class TodoList extends UIElement {
-	connectedCallback() {
-		this.set('filter', 'all') // set initial filter
-		this.#updateList()
-
-		// Event listener and attribute on own element
-		this.self
-			.map(on('click', e => {
-				if (e.target.localName === 'button') this.removeItem(e.target)
-			}))
-			.forEach(setAttribute('filter'))
-
-		// Update count on each change
-		this.set('count', () => {
-			const tasks = this.get('tasks').map(el => el.signal('checked'))
-			const completed = tasks.filter(fn => fn()).length
-			const total = tasks.length
-			return {
-				active: total - completed,
-				completed,
-				total
-			}
-		})
-	}
-
-	addItem = task => {
-		const template = this.querySelector('template').content.cloneNode(true)
-		template.querySelector('span').textContent = task
-		this.querySelector('ol').appendChild(template)
-		this.#updateList()
-	}
-
-	removeItem = element => {
-		element.closest('li').remove()
-		this.#updateList()
-	}
-
-	clearCompleted = () => {
-		this.get('tasks')
-			.filter(el => el.get('checked'))
-			.forEach(el => el.parentElement.remove())
-		this.#updateList()
-	}
-
-	#updateList() {
-		this.set('tasks', Array.from(this.querySelectorAll('input-checkbox')))
-	}
-
-}
-TodoList.define('todo-list')
+var Se=(e)=>typeof e==="function",rt=(e)=>Se(e)&&/^async\s+/.test(e.toString()),ye=(e)=>(t)=>t instanceof e,Oe=ye(Error),st=ye(Promise),$e="Computed",ve=(e,t)=>{t=t??rt(e);let a=[],n,r=null,d=!0,p=()=>{if(d=!0,t)Re(a)},f={[Symbol.toStringTag]:$e,get:()=>{if(t)Ce(a);if(!t||d)De(()=>{let b=(S)=>{n=S,d=!1,r=null},A=(S)=>{r=Oe(S)?S:new Error(`Computed function failed: ${S}`)};try{let S=e(n);st(S)?S.then(b).catch(A):b(S)}catch(S){A(S)}},p);if(Oe(r))throw r;return n},map:(b)=>ve(()=>b(f.get()))};return f},it=(e)=>!!e&&typeof e==="object"&&e[Symbol.toStringTag]===$e,H,lt=!1,ot=[],fe=(e)=>me(e)||it(e),Ne=(e,t=!1)=>fe(e)?e:Se(e)?ve(e,t):ut(e),Ce=(e)=>{if(H&&!e.includes(H))e.push(H)},Re=(e)=>e.forEach((t)=>lt?ot.push(t):t()),De=(e,t)=>{let a=H;H=t,e(),H=a};var pe=Symbol();class ke{$;watchers=[];constructor(e){this.value=e}get(){return Ce(this.watchers),this.value}set(e){if(pe!==e){let t=Se(e)?e(this.value):e;if(Object.is(this.value,t))return;this.value=t}if(Re(this.watchers),pe===e)this.watchers=[]}map(e){return ve(()=>e(this.get()))}}var ut=(e)=>new ke(e),me=ye(ke),B=(e)=>{let t=()=>De(()=>{try{e()}catch(a){console.error(a)}},t);t()};if(!("requestAnimationFrame"in globalThis))globalThis.requestAnimationFrame=(e)=>setTimeout(e,16);var re=new Map,Y=[],se,Le=()=>{se=null,Y.forEach((e)=>e()),Y=[],re.clear()},ct=()=>{if(se)cancelAnimationFrame(se);se=requestAnimationFrame(Le)};queueMicrotask(Le);var L=(e,t)=>new Promise((a,n)=>{let r=()=>{try{a(e())}catch(d){n(d)}};if(t){let[d,p]=t;if(!re.has(d))re.set(d,new Map);let f=re.get(d);if(f.has(p)){let b=Y.indexOf(e);if(b>-1)Y.splice(b,1)}f.set(p,r)}Y.push(r),ct()});var dt=(e)=>e.nodeType===Node.COMMENT_NODE,gt=(e)=>!/^on/i.test(e),ht=(e)=>{if(/^(mailto|tel):/i.test(e))return!0;if(e.includes("://"))try{let t=new URL(e,window.location.origin);return!["http:","https:","ftp:"].includes(t.protocol)}catch(t){return!0}return!0},ft=(e,t,a)=>{if(!gt(t))throw new Error(`Unsafe attribute: ${t}`);if(a=String(a).trim(),!ht(a))throw new Error(`Unsafe URL for ${t}: ${a}`);e.setAttribute(t,a)};var pt=(e,t)=>L(()=>{return Array.from(e.childNodes).filter((a)=>!dt(a)).forEach((a)=>a.remove()),e.append(document.createTextNode(t)),e},[e,"t"]),mt=(e,t,a)=>L(()=>{return ft(e,t,a),e},[e,`a:${t}`]),bt=(e,t)=>L(()=>{return e.removeAttribute(t),e},[e,`a:${t}`]),St=(e,t,a)=>L(()=>{return e.toggleAttribute(t,a),e},[e,`a:${t}`]),yt=(e,t,a)=>L(()=>{return e.classList.toggle(t,a),e},[e,`c:${t}`]);var D=(e)=>typeof e==="function",Ge=(e)=>!!e&&typeof e==="object",vt=(e)=>typeof e==="number",le=(e)=>typeof e==="string",kt=(e)=>typeof e==="symbol",Ft=(e)=>le(e)||kt(e)||vt(e),x=!1,wt="debug";var At="warn",X="error",xt=(e)=>e?`#${e}`:"",Et=(e)=>e.length?`.${Array.from(e).join(".")}`:"",J=(e)=>`<${e.localName}${xt(e.id)}${Et(e.classList)}>`,R=(e)=>le(e)?`"${e}"`:Ge(e)?JSON.stringify(e):String(e),E=(e,t,a=wt)=>{if(x||[X,At].includes(a))console[a](t,e);return e},_t=(e)=>D(e)&&e.length===2,he=(e,t,a=0)=>_t(e)?e(t,a):e;class ie{$;A;constructor(e,t=[e]){this.host=e,this.targets=t}on(e,t){return this.targets.forEach((a,n)=>a.addEventListener(e,he(t,a,n))),this}off(e,t){return this.targets.forEach((a,n)=>a.removeEventListener(e,he(t,a,n))),this}pass(e){return this.targets.forEach(async(t,a)=>{if(await y.registry.whenDefined(t.localName),t instanceof y)Object.entries(e).forEach(([n,r])=>{let d=he(r,t,a),p=Ft(d)?this.host.signals.get(d):Ne(d,!0);if(p)t.set(n,p);else E(r,`Invalid source for state ${R(n)}`,X)});else E(t,"Target is not a UIElement",X)}),this}sync(...e){return this.targets.forEach((t,a)=>e.forEach((n)=>n(this.host,t,a))),this}}var Ot=(e,t)=>{if(t==null)return;let a=e(t);return Number.isFinite(a)?a:void 0},be=(e,t,a,n=void 0)=>{let r=e.constructor.states[t];return D(r)&&!!r.length?r(a,e,n):a},I=(e)=>e!=null,je=(e)=>Ot(parseInt,e);var Te="context-request";class Pe extends Event{$;A;B;constructor(e,t,a=!1){super(Te,{bubbles:!0,composed:!0});this.context=e,this.callback=t,this.subscribe=a}}var $t=(e)=>{let t=e.constructor,a=t.consumedContexts||[];setTimeout(()=>{for(let r of a)e.dispatchEvent(new Pe(r,(d)=>e.set(String(r),d)))});let n=t.providedContexts||[];if(!n.length)return!1;return e.addEventListener(Te,(r)=>{let{context:d,callback:p}=r;if(!n.includes(d)||!D(p))return;r.stopPropagation(),p(e.signals.get(String(d)))}),!0};class y extends HTMLElement{static registry=customElements;static states={};static observedAttributes;static consumedContexts;static providedContexts;static define(e){try{if(y.registry.define(e,this),x)E(e,"Registered custom element")}catch(t){E(t,`Failed to register custom element ${e}`,X)}}signals=new Map;self=new ie(this);root=this.shadowRoot||this;debug=!1;attributeChangedCallback(e,t,a){if(a===t)return;if(x&&this.debug)E(`${R(t)} => ${R(a)}`,`Attribute "${e}" of ${J(this)} changed`);this.set(e,be(this,e,a,t))}connectedCallback(){if(x){if(this.debug=this.hasAttribute("debug"),this.debug)E(this,"Connected")}Object.entries(this.constructor.states).forEach(([e,t])=>{let a=D(t)?be(this,e,this.getAttribute(e)??void 0,void 0):t;this.set(e,a,!1)}),$t(this)}disconnectedCallback(){if(x&&this.debug)E(this,"Disconnected")}adoptedCallback(){if(x&&this.debug)E(this,"Adopted")}has(e){return this.signals.has(e)}get(e){let t=(n)=>!Ge(n)?n:D(n)?t(n()):fe(n)?t(n.get()):n,a=t(this.signals.get(e));if(x&&this.debug)E(a,`Get current value of state ${R(e)} in ${J(this)}`);return a}set(e,t,a=!0){let n;if(!this.signals.has(e)){if(x&&this.debug)n="Create";this.signals.set(e,Ne(t,!0))}else if(a){let r=this.signals.get(e);if(fe(t)){if(x&&this.debug)n="Replace";if(this.signals.set(e,t),me(r))r.set(pe)}else if(me(r)){if(x&&this.debug)n="Update";r.set(t)}else{E(t,`Computed state ${R(e)} in ${J(this)} cannot be set`,X);return}}else return;if(x&&this.debug)E(t,`${n} state ${R(e)} in ${J(this)}`)}delete(e){if(x&&this.debug)E(e,`Delete state ${R(e)} from ${J(this)}`);return this.signals.delete(e)}first(e){let t=this.root.querySelector(e);return new ie(this,t?[t]:[])}all(e){return new ie(this,Array.from(this.root.querySelectorAll(e)))}}var Q=(e,t)=>(a,n)=>{let{read:r,update:d}=t,p=r(n);if(!D(e)){let f=le(e)&&le(p)?be(a,e,p):p;a.set(e,f,!1)}B(()=>{let f=r(n),b=D(e)?e(f):a.get(e);if(!Object.is(b,f)){if(b===null&&t.delete)t.delete(n);else if(b==null&&p)d(n,p);else if(b!=null)d(n,b)}})};var $=(e)=>Q(e,{read:(t)=>t.textContent,update:(t,a)=>pt(t,a)}),F=(e,t=e)=>Q(t,{read:(a)=>a[e],update:(a,n)=>a[e]=n}),W=(e,t=e)=>Q(t,{read:(a)=>a.getAttribute(e),update:(a,n)=>mt(a,e,n),delete:(a)=>bt(a,e)}),U=(e,t=e)=>Q(t,{read:(a)=>a.hasAttribute(e),update:(a,n)=>St(a,e,n)}),M=(e,t=e)=>Q(t,{read:(a)=>a.classList.contains(e),update:(a,n)=>yt(a,e,n)});var Nt=typeof window!="undefined"?window:typeof WorkerGlobalScope!="undefined"&&self instanceof WorkerGlobalScope?self:{},g=function(e){var t=/(?:^|\s)lang(?:uage)?-([\w-]+)(?=\s|$)/i,a=0,n={},r={manual:e.Prism&&e.Prism.manual,disableWorkerMessageHandler:e.Prism&&e.Prism.disableWorkerMessageHandler,util:{encode:function i(s){return s instanceof d?new d(s.type,i(s.content),s.alias):Array.isArray(s)?s.map(i):s.replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/\u00a0/g," ")},type:function(i){return Object.prototype.toString.call(i).slice(8,-1)},objId:function(i){return i.__id||Object.defineProperty(i,"__id",{value:++a}),i.__id},clone:function i(s,u){var l,o;switch(u=u||{},r.util.type(s)){case"Object":if(o=r.util.objId(s),u[o])return u[o];for(var c in l={},u[o]=l,s)s.hasOwnProperty(c)&&(l[c]=i(s[c],u));return l;case"Array":return o=r.util.objId(s),u[o]?u[o]:(l=[],u[o]=l,s.forEach(function(h,m){l[m]=i(h,u)}),l);default:return s}},getLanguage:function(i){for(;i;){var s=t.exec(i.className);if(s)return s[1].toLowerCase();i=i.parentElement}return"none"},setLanguage:function(i,s){i.className=i.className.replace(RegExp(t,"gi"),""),i.classList.add("language-"+s)},currentScript:function(){if(typeof document=="undefined")return null;if("currentScript"in document)return document.currentScript;try{throw new Error}catch(l){var i=(/at [^(\r\n]*\((.*):[^:]+:[^:]+\)$/i.exec(l.stack)||[])[1];if(i){var s=document.getElementsByTagName("script");for(var u in s)if(s[u].src==i)return s[u]}return null}},isActive:function(i,s,u){for(var l="no-"+s;i;){var o=i.classList;if(o.contains(s))return!0;if(o.contains(l))return!1;i=i.parentElement}return!!u}},languages:{plain:n,plaintext:n,text:n,txt:n,extend:function(i,s){var u=r.util.clone(r.languages[i]);for(var l in s)u[l]=s[l];return u},insertBefore:function(i,s,u,l){var o=(l=l||r.languages)[i],c={};for(var h in o)if(o.hasOwnProperty(h)){if(h==s)for(var m in u)u.hasOwnProperty(m)&&(c[m]=u[m]);u.hasOwnProperty(h)||(c[h]=o[h])}var k=l[i];return l[i]=c,r.languages.DFS(r.languages,function(v,j){j===k&&v!=i&&(this[v]=c)}),c},DFS:function i(s,u,l,o){o=o||{};var c=r.util.objId;for(var h in s)if(s.hasOwnProperty(h)){u.call(s,h,s[h],l||h);var m=s[h],k=r.util.type(m);k!=="Object"||o[c(m)]?k!=="Array"||o[c(m)]||(o[c(m)]=!0,i(m,u,h,o)):(o[c(m)]=!0,i(m,u,null,o))}}},plugins:{},highlightAll:function(i,s){r.highlightAllUnder(document,i,s)},highlightAllUnder:function(i,s,u){var l={callback:u,container:i,selector:'code[class*="language-"], [class*="language-"] code, code[class*="lang-"], [class*="lang-"] code'};r.hooks.run("before-highlightall",l),l.elements=Array.prototype.slice.apply(l.container.querySelectorAll(l.selector)),r.hooks.run("before-all-elements-highlight",l);for(var o,c=0;o=l.elements[c++];)r.highlightElement(o,s===!0,l.callback)},highlightElement:function(i,s,u){var l=r.util.getLanguage(i),o=r.languages[l];r.util.setLanguage(i,l);var c=i.parentElement;c&&c.nodeName.toLowerCase()==="pre"&&r.util.setLanguage(c,l);var h={element:i,language:l,grammar:o,code:i.textContent};function m(v){h.highlightedCode=v,r.hooks.run("before-insert",h),h.element.innerHTML=h.highlightedCode,r.hooks.run("after-highlight",h),r.hooks.run("complete",h),u&&u.call(h.element)}if(r.hooks.run("before-sanity-check",h),(c=h.element.parentElement)&&c.nodeName.toLowerCase()==="pre"&&!c.hasAttribute("tabindex")&&c.setAttribute("tabindex","0"),!h.code)return r.hooks.run("complete",h),void(u&&u.call(h.element));if(r.hooks.run("before-highlight",h),h.grammar)if(s&&e.Worker){var k=new Worker(r.filename);k.onmessage=function(v){m(v.data)},k.postMessage(JSON.stringify({language:h.language,code:h.code,immediateClose:!0}))}else m(r.highlight(h.code,h.grammar,h.language));else m(r.util.encode(h.code))},highlight:function(i,s,u){var l={code:i,grammar:s,language:u};if(r.hooks.run("before-tokenize",l),!l.grammar)throw new Error('The language "'+l.language+'" has no grammar.');return l.tokens=r.tokenize(l.code,l.grammar),r.hooks.run("after-tokenize",l),d.stringify(r.util.encode(l.tokens),l.language)},tokenize:function(i,s){var u=s.rest;if(u){for(var l in u)s[l]=u[l];delete s.rest}var o=new b;return A(o,o.head,i),f(i,o,s,o.head,0),function(c){for(var h=[],m=c.head.next;m!==c.tail;)h.push(m.value),m=m.next;return h}(o)},hooks:{all:{},add:function(i,s){var u=r.hooks.all;u[i]=u[i]||[],u[i].push(s)},run:function(i,s){var u=r.hooks.all[i];if(u&&u.length)for(var l,o=0;l=u[o++];)l(s)}},Token:d};function d(i,s,u,l){this.type=i,this.content=s,this.alias=u,this.length=0|(l||"").length}function p(i,s,u,l){i.lastIndex=s;var o=i.exec(u);if(o&&l&&o[1]){var c=o[1].length;o.index+=c,o[0]=o[0].slice(c)}return o}function f(i,s,u,l,o,c){for(var h in u)if(u.hasOwnProperty(h)&&u[h]){var m=u[h];m=Array.isArray(m)?m:[m];for(var k=0;k<m.length;++k){if(c&&c.cause==h+","+k)return;var v=m[k],j=v.inside,T=!!v.lookbehind,K=!!v.greedy,V=v.alias;if(K&&!v.pattern.global){var at=v.pattern.toString().match(/[imsuy]*$/)[0];v.pattern=RegExp(v.pattern.source,at+"g")}for(var Ee=v.pattern||v,w=l.next,_=o;w!==s.tail&&!(c&&_>=c.reach);_+=w.value.length,w=w.next){var P=w.value;if(s.length>i.length)return;if(!(P instanceof d)){var O,ee=1;if(K){if(!(O=p(Ee,_,i,T))||O.index>=i.length)break;var te=O.index,nt=O.index+O[0].length,C=_;for(C+=w.value.length;te>=C;)C+=(w=w.next).value.length;if(_=C-=w.value.length,w.value instanceof d)continue;for(var q=w;q!==s.tail&&(C<nt||typeof q.value=="string");q=q.next)ee++,C+=q.value.length;ee--,P=i.slice(_,C),O.index-=_}else if(!(O=p(Ee,0,P,T)))continue;te=O.index;var ae=O[0],ce=P.slice(0,te),_e=P.slice(te+ae.length),de=_+P.length;c&&de>c.reach&&(c.reach=de);var ne=w.prev;if(ce&&(ne=A(s,ne,ce),_+=ce.length),S(s,ne,ee),w=A(s,ne,new d(h,j?r.tokenize(ae,j):ae,V,ae)),_e&&A(s,w,_e),ee>1){var ge={cause:h+","+k,reach:de};f(i,s,u,w.prev,_,ge),c&&ge.reach>c.reach&&(c.reach=ge.reach)}}}}}}function b(){var i={value:null,prev:null,next:null},s={value:null,prev:i,next:null};i.next=s,this.head=i,this.tail=s,this.length=0}function A(i,s,u){var l=s.next,o={value:u,prev:s,next:l};return s.next=o,l.prev=o,i.length++,o}function S(i,s,u){for(var l=s.next,o=0;o<u&&l!==i.tail;o++)l=l.next;s.next=l,l.prev=s,i.length-=o}if(e.Prism=r,d.stringify=function i(s,u){if(typeof s=="string")return s;if(Array.isArray(s)){var l="";return s.forEach(function(k){l+=i(k,u)}),l}var o={type:s.type,content:i(s.content,u),tag:"span",classes:["token",s.type],attributes:{},language:u},c=s.alias;c&&(Array.isArray(c)?Array.prototype.push.apply(o.classes,c):o.classes.push(c)),r.hooks.run("wrap",o);var h="";for(var m in o.attributes)h+=" "+m+'="'+(o.attributes[m]||"").replace(/"/g,"&quot;")+'"';return"<"+o.tag+' class="'+o.classes.join(" ")+'"'+h+">"+o.content+"</"+o.tag+">"},!e.document)return e.addEventListener?(r.disableWorkerMessageHandler||e.addEventListener("message",function(i){var s=JSON.parse(i.data),u=s.language,l=s.code,o=s.immediateClose;e.postMessage(r.highlight(l,r.languages[u],u)),o&&e.close()},!1),r):r;var N=r.util.currentScript();function G(){r.manual||r.highlightAll()}if(N&&(r.filename=N.src,N.hasAttribute("data-manual")&&(r.manual=!0)),!r.manual){var Z=document.readyState;Z==="loading"||Z==="interactive"&&N&&N.defer?document.addEventListener("DOMContentLoaded",G):window.requestAnimationFrame?window.requestAnimationFrame(G):window.setTimeout(G,16)}return r}(Nt);typeof Fe!="undefined"&&Fe.exports&&(Fe.exports=g),typeof global!="undefined"&&(global.Prism=g);g.languages.markup={comment:{pattern:/<!--(?:(?!<!--)[\s\S])*?-->/,greedy:!0},prolog:{pattern:/<\?[\s\S]+?\?>/,greedy:!0},doctype:{pattern:/<!DOCTYPE(?:[^>"'[\]]|"[^"]*"|'[^']*')+(?:\[(?:[^<"'\]]|"[^"]*"|'[^']*'|<(?!!--)|<!--(?:[^-]|-(?!->))*-->)*\]\s*)?>/i,greedy:!0,inside:{"internal-subset":{pattern:/(^[^\[]*\[)[\s\S]+(?=\]>$)/,lookbehind:!0,greedy:!0,inside:null},string:{pattern:/"[^"]*"|'[^']*'/,greedy:!0},punctuation:/^<!|>$|[[\]]/,"doctype-tag":/^DOCTYPE/i,name:/[^\s<>'"]+/}},cdata:{pattern:/<!\[CDATA\[[\s\S]*?\]\]>/i,greedy:!0},tag:{pattern:/<\/?(?!\d)[^\s>\/=$<%]+(?:\s(?:\s*[^\s>\/=]+(?:\s*=\s*(?:"[^"]*"|'[^']*'|[^\s'">=]+(?=[\s>]))|(?=[\s/>])))+)?\s*\/?>/,greedy:!0,inside:{tag:{pattern:/^<\/?[^\s>\/]+/,inside:{punctuation:/^<\/?/,namespace:/^[^\s>\/:]+:/}},"special-attr":[],"attr-value":{pattern:/=\s*(?:"[^"]*"|'[^']*'|[^\s'">=]+)/,inside:{punctuation:[{pattern:/^=/,alias:"attr-equals"},{pattern:/^(\s*)["']|["']$/,lookbehind:!0}]}},punctuation:/\/?>/,"attr-name":{pattern:/[^\s>\/]+/,inside:{namespace:/^[^\s>\/:]+:/}}}},entity:[{pattern:/&[\da-z]{1,8};/i,alias:"named-entity"},/&#x?[\da-f]{1,8};/i]},g.languages.markup.tag.inside["attr-value"].inside.entity=g.languages.markup.entity,g.languages.markup.doctype.inside["internal-subset"].inside=g.languages.markup,g.hooks.add("wrap",function(e){e.type==="entity"&&(e.attributes.title=e.content.replace(/&amp;/,"&"))}),Object.defineProperty(g.languages.markup.tag,"addInlined",{value:function(e,t){var a={};a["language-"+t]={pattern:/(^<!\[CDATA\[)[\s\S]+?(?=\]\]>$)/i,lookbehind:!0,inside:g.languages[t]},a.cdata=/^<!\[CDATA\[|\]\]>$/i;var n={"included-cdata":{pattern:/<!\[CDATA\[[\s\S]*?\]\]>/i,inside:a}};n["language-"+t]={pattern:/[\s\S]+/,inside:g.languages[t]};var r={};r[e]={pattern:RegExp("(<__[^>]*>)(?:<!\\[CDATA\\[(?:[^\\]]|\\](?!\\]>))*\\]\\]>|(?!<!\\[CDATA\\[)[^])*?(?=</__>)".replace(/__/g,function(){return e}),"i"),lookbehind:!0,greedy:!0,inside:n},g.languages.insertBefore("markup","cdata",r)}}),Object.defineProperty(g.languages.markup.tag,"addAttribute",{value:function(e,t){g.languages.markup.tag.inside["special-attr"].push({pattern:RegExp("(^|[\"'\\s])(?:"+e+")\\s*=\\s*(?:\"[^\"]*\"|'[^']*'|[^\\s'\">=]+(?=[\\s>]))","i"),lookbehind:!0,inside:{"attr-name":/^[^\s=]+/,"attr-value":{pattern:/=[\s\S]+/,inside:{value:{pattern:/(^=\s*(["']|(?!["'])))\S[\s\S]*(?=\2$)/,lookbehind:!0,alias:[t,"language-"+t],inside:g.languages[t]},punctuation:[{pattern:/^=/,alias:"attr-equals"},/"|'/]}}}})}}),g.languages.html=g.languages.markup,g.languages.mathml=g.languages.markup,g.languages.svg=g.languages.markup,g.languages.xml=g.languages.extend("markup",{}),g.languages.ssml=g.languages.xml,g.languages.atom=g.languages.xml,g.languages.rss=g.languages.xml;(function(e){var t=/(?:"(?:\\(?:\r\n|[\s\S])|[^"\\\r\n])*"|'(?:\\(?:\r\n|[\s\S])|[^'\\\r\n])*')/;e.languages.css={comment:/\/\*[\s\S]*?\*\//,atrule:{pattern:RegExp("@[\\w-](?:[^;{\\s\"']|\\s+(?!\\s)|"+t.source+")*?(?:;|(?=\\s*\\{))"),inside:{rule:/^@[\w-]+/,"selector-function-argument":{pattern:/(\bselector\s*\(\s*(?![\s)]))(?:[^()\s]|\s+(?![\s)])|\((?:[^()]|\([^()]*\))*\))+(?=\s*\))/,lookbehind:!0,alias:"selector"},keyword:{pattern:/(^|[^\w-])(?:and|not|only|or)(?![\w-])/,lookbehind:!0}}},url:{pattern:RegExp("\\burl\\((?:"+t.source+"|(?:[^\\\\\r\n()\"']|\\\\[^])*)\\)","i"),greedy:!0,inside:{function:/^url/i,punctuation:/^\(|\)$/,string:{pattern:RegExp("^"+t.source+"$"),alias:"url"}}},selector:{pattern:RegExp("(^|[{}\\s])[^{}\\s](?:[^{};\"'\\s]|\\s+(?![\\s{])|"+t.source+")*(?=\\s*\\{)"),lookbehind:!0},string:{pattern:t,greedy:!0},property:{pattern:/(^|[^-\w\xA0-\uFFFF])(?!\s)[-_a-z\xA0-\uFFFF](?:(?!\s)[-\w\xA0-\uFFFF])*(?=\s*:)/i,lookbehind:!0},important:/!important\b/i,function:{pattern:/(^|[^-a-z0-9])[-a-z0-9]+(?=\()/i,lookbehind:!0},punctuation:/[(){};:,]/},e.languages.css.atrule.inside.rest=e.languages.css;var a=e.languages.markup;a&&(a.tag.addInlined("style","css"),a.tag.addAttribute("style","css"))})(g);g.languages.clike={comment:[{pattern:/(^|[^\\])\/\*[\s\S]*?(?:\*\/|$)/,lookbehind:!0,greedy:!0},{pattern:/(^|[^\\:])\/\/.*/,lookbehind:!0,greedy:!0}],string:{pattern:/(["'])(?:\\(?:\r\n|[\s\S])|(?!\1)[^\\\r\n])*\1/,greedy:!0},"class-name":{pattern:/(\b(?:class|extends|implements|instanceof|interface|new|trait)\s+|\bcatch\s+\()[\w.\\]+/i,lookbehind:!0,inside:{punctuation:/[.\\]/}},keyword:/\b(?:break|catch|continue|do|else|finally|for|function|if|in|instanceof|new|null|return|throw|try|while)\b/,boolean:/\b(?:false|true)\b/,function:/\b\w+(?=\()/,number:/\b0x[\da-f]+\b|(?:\b\d+(?:\.\d*)?|\B\.\d+)(?:e[+-]?\d+)?/i,operator:/[<>]=?|[!=]=?=?|--?|\+\+?|&&?|\|\|?|[?*/~^%]/,punctuation:/[{}[\];(),.:]/};g.languages.javascript=g.languages.extend("clike",{"class-name":[g.languages.clike["class-name"],{pattern:/(^|[^$\w\xA0-\uFFFF])(?!\s)[_$A-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\.(?:constructor|prototype))/,lookbehind:!0}],keyword:[{pattern:/((?:^|\})\s*)catch\b/,lookbehind:!0},{pattern:/(^|[^.]|\.\.\.\s*)\b(?:as|assert(?=\s*\{)|async(?=\s*(?:function\b|\(|[$\w\xA0-\uFFFF]|$))|await|break|case|class|const|continue|debugger|default|delete|do|else|enum|export|extends|finally(?=\s*(?:\{|$))|for|from(?=\s*(?:['"]|$))|function|(?:get|set)(?=\s*(?:[#\[$\w\xA0-\uFFFF]|$))|if|implements|import|in|instanceof|interface|let|new|null|of|package|private|protected|public|return|static|super|switch|this|throw|try|typeof|undefined|var|void|while|with|yield)\b/,lookbehind:!0}],function:/#?(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\s*(?:\.\s*(?:apply|bind|call)\s*)?\()/,number:{pattern:RegExp("(^|[^\\w$])(?:NaN|Infinity|0[bB][01]+(?:_[01]+)*n?|0[oO][0-7]+(?:_[0-7]+)*n?|0[xX][\\dA-Fa-f]+(?:_[\\dA-Fa-f]+)*n?|\\d+(?:_\\d+)*n|(?:\\d+(?:_\\d+)*(?:\\.(?:\\d+(?:_\\d+)*)?)?|\\.\\d+(?:_\\d+)*)(?:[Ee][+-]?\\d+(?:_\\d+)*)?)(?![\\w$])"),lookbehind:!0},operator:/--|\+\+|\*\*=?|=>|&&=?|\|\|=?|[!=]==|<<=?|>>>?=?|[-+*/%&|^!=<>]=?|\.{3}|\?\?=?|\?\.?|[~:]/}),g.languages.javascript["class-name"][0].pattern=/(\b(?:class|extends|implements|instanceof|interface|new)\s+)[\w.\\]+/,g.languages.insertBefore("javascript","keyword",{regex:{pattern:RegExp("((?:^|[^$\\w\\xA0-\\uFFFF.\"'\\])\\s]|\\b(?:return|yield))\\s*)/(?:(?:\\[(?:[^\\]\\\\\r\n]|\\\\.)*\\]|\\\\.|[^/\\\\\\[\r\n])+/[dgimyus]{0,7}|(?:\\[(?:[^[\\]\\\\\r\n]|\\\\.|\\[(?:[^[\\]\\\\\r\n]|\\\\.|\\[(?:[^[\\]\\\\\r\n]|\\\\.)*\\])*\\])*\\]|\\\\.|[^/\\\\\\[\r\n])+/[dgimyus]{0,7}v[dgimyus]{0,7})(?=(?:\\s|/\\*(?:[^*]|\\*(?!/))*\\*/)*(?:$|[\r\n,.;:})\\]]|//))"),lookbehind:!0,greedy:!0,inside:{"regex-source":{pattern:/^(\/)[\s\S]+(?=\/[a-z]*$)/,lookbehind:!0,alias:"language-regex",inside:g.languages.regex},"regex-delimiter":/^\/|\/$/,"regex-flags":/^[a-z]+$/}},"function-variable":{pattern:/#?(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\s*[=:]\s*(?:async\s*)?(?:\bfunction\b|(?:\((?:[^()]|\([^()]*\))*\)|(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*)\s*=>))/,alias:"function"},parameter:[{pattern:/(function(?:\s+(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*)?\s*\(\s*)(?!\s)(?:[^()\s]|\s+(?![\s)])|\([^()]*\))+(?=\s*\))/,lookbehind:!0,inside:g.languages.javascript},{pattern:/(^|[^$\w\xA0-\uFFFF])(?!\s)[_$a-z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\s*=>)/i,lookbehind:!0,inside:g.languages.javascript},{pattern:/(\(\s*)(?!\s)(?:[^()\s]|\s+(?![\s)])|\([^()]*\))+(?=\s*\)\s*=>)/,lookbehind:!0,inside:g.languages.javascript},{pattern:/((?:\b|\s|^)(?!(?:as|async|await|break|case|catch|class|const|continue|debugger|default|delete|do|else|enum|export|extends|finally|for|from|function|get|if|implements|import|in|instanceof|interface|let|new|null|of|package|private|protected|public|return|set|static|super|switch|this|throw|try|typeof|undefined|var|void|while|with|yield)(?![$\w\xA0-\uFFFF]))(?:(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*\s*)\(\s*|\]\s*\(\s*)(?!\s)(?:[^()\s]|\s+(?![\s)])|\([^()]*\))+(?=\s*\)\s*\{)/,lookbehind:!0,inside:g.languages.javascript}],constant:/\b[A-Z](?:[A-Z_]|\dx?)*\b/}),g.languages.insertBefore("javascript","string",{hashbang:{pattern:/^#!.*/,greedy:!0,alias:"comment"},"template-string":{pattern:/`(?:\\[\s\S]|\$\{(?:[^{}]|\{(?:[^{}]|\{[^}]*\})*\})+\}|(?!\$\{)[^\\`])*`/,greedy:!0,inside:{"template-punctuation":{pattern:/^`|`$/,alias:"string"},interpolation:{pattern:/((?:^|[^\\])(?:\\{2})*)\$\{(?:[^{}]|\{(?:[^{}]|\{[^}]*\})*\})+\}/,lookbehind:!0,inside:{"interpolation-punctuation":{pattern:/^\$\{|\}$/,alias:"punctuation"},rest:g.languages.javascript}},string:/[\s\S]+/}},"string-property":{pattern:/((?:^|[,{])[ \t]*)(["'])(?:\\(?:\r\n|[\s\S])|(?!\2)[^\\\r\n])*\2(?=\s*:)/m,lookbehind:!0,greedy:!0,alias:"property"}}),g.languages.insertBefore("javascript","operator",{"literal-property":{pattern:/((?:^|[,{])[ \t]*)(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\s*:)/m,lookbehind:!0,alias:"property"}}),g.languages.markup&&(g.languages.markup.tag.addInlined("script","javascript"),g.languages.markup.tag.addAttribute("on(?:abort|blur|change|click|composition(?:end|start|update)|dblclick|error|focus(?:in|out)?|key(?:down|up)|load|mouse(?:down|enter|leave|move|out|over|up)|reset|resize|scroll|select|slotchange|submit|unload|wheel)","javascript")),g.languages.js=g.languages.javascript;(function(e){var t="\\b(?:BASH|BASHOPTS|BASH_ALIASES|BASH_ARGC|BASH_ARGV|BASH_CMDS|BASH_COMPLETION_COMPAT_DIR|BASH_LINENO|BASH_REMATCH|BASH_SOURCE|BASH_VERSINFO|BASH_VERSION|COLORTERM|COLUMNS|COMP_WORDBREAKS|DBUS_SESSION_BUS_ADDRESS|DEFAULTS_PATH|DESKTOP_SESSION|DIRSTACK|DISPLAY|EUID|GDMSESSION|GDM_LANG|GNOME_KEYRING_CONTROL|GNOME_KEYRING_PID|GPG_AGENT_INFO|GROUPS|HISTCONTROL|HISTFILE|HISTFILESIZE|HISTSIZE|HOME|HOSTNAME|HOSTTYPE|IFS|INSTANCE|JOB|LANG|LANGUAGE|LC_ADDRESS|LC_ALL|LC_IDENTIFICATION|LC_MEASUREMENT|LC_MONETARY|LC_NAME|LC_NUMERIC|LC_PAPER|LC_TELEPHONE|LC_TIME|LESSCLOSE|LESSOPEN|LINES|LOGNAME|LS_COLORS|MACHTYPE|MAILCHECK|MANDATORY_PATH|NO_AT_BRIDGE|OLDPWD|OPTERR|OPTIND|ORBIT_SOCKETDIR|OSTYPE|PAPERSIZE|PATH|PIPESTATUS|PPID|PS1|PS2|PS3|PS4|PWD|RANDOM|REPLY|SECONDS|SELINUX_INIT|SESSION|SESSIONTYPE|SESSION_MANAGER|SHELL|SHELLOPTS|SHLVL|SSH_AUTH_SOCK|TERM|UID|UPSTART_EVENTS|UPSTART_INSTANCE|UPSTART_JOB|UPSTART_SESSION|USER|WINDOWID|XAUTHORITY|XDG_CONFIG_DIRS|XDG_CURRENT_DESKTOP|XDG_DATA_DIRS|XDG_GREETER_DATA_DIR|XDG_MENU_PREFIX|XDG_RUNTIME_DIR|XDG_SEAT|XDG_SEAT_PATH|XDG_SESSION_DESKTOP|XDG_SESSION_ID|XDG_SESSION_PATH|XDG_SESSION_TYPE|XDG_VTNR|XMODIFIERS)\\b",a={pattern:/(^(["']?)\w+\2)[ \t]+\S.*/,lookbehind:!0,alias:"punctuation",inside:null},n={bash:a,environment:{pattern:RegExp("\\$"+t),alias:"constant"},variable:[{pattern:/\$?\(\([\s\S]+?\)\)/,greedy:!0,inside:{variable:[{pattern:/(^\$\(\([\s\S]+)\)\)/,lookbehind:!0},/^\$\(\(/],number:/\b0x[\dA-Fa-f]+\b|(?:\b\d+(?:\.\d*)?|\B\.\d+)(?:[Ee]-?\d+)?/,operator:/--|\+\+|\*\*=?|<<=?|>>=?|&&|\|\||[=!+\-*/%<>^&|]=?|[?~:]/,punctuation:/\(\(?|\)\)?|,|;/}},{pattern:/\$\((?:\([^)]+\)|[^()])+\)|`[^`]+`/,greedy:!0,inside:{variable:/^\$\(|^`|\)$|`$/}},{pattern:/\$\{[^}]+\}/,greedy:!0,inside:{operator:/:[-=?+]?|[!\/]|##?|%%?|\^\^?|,,?/,punctuation:/[\[\]]/,environment:{pattern:RegExp("(\\{)"+t),lookbehind:!0,alias:"constant"}}},/\$(?:\w+|[#?*!@$])/],entity:/\\(?:[abceEfnrtv\\"]|O?[0-7]{1,3}|U[0-9a-fA-F]{8}|u[0-9a-fA-F]{4}|x[0-9a-fA-F]{1,2})/};e.languages.bash={shebang:{pattern:/^#!\s*\/.*/,alias:"important"},comment:{pattern:/(^|[^"{\\$])#.*/,lookbehind:!0},"function-name":[{pattern:/(\bfunction\s+)[\w-]+(?=(?:\s*\(?:\s*\))?\s*\{)/,lookbehind:!0,alias:"function"},{pattern:/\b[\w-]+(?=\s*\(\s*\)\s*\{)/,alias:"function"}],"for-or-select":{pattern:/(\b(?:for|select)\s+)\w+(?=\s+in\s)/,alias:"variable",lookbehind:!0},"assign-left":{pattern:/(^|[\s;|&]|[<>]\()\w+(?:\.\w+)*(?=\+?=)/,inside:{environment:{pattern:RegExp("(^|[\\s;|&]|[<>]\\()"+t),lookbehind:!0,alias:"constant"}},alias:"variable",lookbehind:!0},parameter:{pattern:/(^|\s)-{1,2}(?:\w+:[+-]?)?\w+(?:\.\w+)*(?=[=\s]|$)/,alias:"variable",lookbehind:!0},string:[{pattern:/((?:^|[^<])<<-?\s*)(\w+)\s[\s\S]*?(?:\r?\n|\r)\2/,lookbehind:!0,greedy:!0,inside:n},{pattern:/((?:^|[^<])<<-?\s*)(["'])(\w+)\2\s[\s\S]*?(?:\r?\n|\r)\3/,lookbehind:!0,greedy:!0,inside:{bash:a}},{pattern:/(^|[^\\](?:\\\\)*)"(?:\\[\s\S]|\$\([^)]+\)|\$(?!\()|`[^`]+`|[^"\\`$])*"/,lookbehind:!0,greedy:!0,inside:n},{pattern:/(^|[^$\\])'[^']*'/,lookbehind:!0,greedy:!0},{pattern:/\$'(?:[^'\\]|\\[\s\S])*'/,greedy:!0,inside:{entity:n.entity}}],environment:{pattern:RegExp("\\$?"+t),alias:"constant"},variable:n.variable,function:{pattern:/(^|[\s;|&]|[<>]\()(?:add|apropos|apt|apt-cache|apt-get|aptitude|aspell|automysqlbackup|awk|basename|bash|bc|bconsole|bg|bzip2|cal|cargo|cat|cfdisk|chgrp|chkconfig|chmod|chown|chroot|cksum|clear|cmp|column|comm|composer|cp|cron|crontab|csplit|curl|cut|date|dc|dd|ddrescue|debootstrap|df|diff|diff3|dig|dir|dircolors|dirname|dirs|dmesg|docker|docker-compose|du|egrep|eject|env|ethtool|expand|expect|expr|fdformat|fdisk|fg|fgrep|file|find|fmt|fold|format|free|fsck|ftp|fuser|gawk|git|gparted|grep|groupadd|groupdel|groupmod|groups|grub-mkconfig|gzip|halt|head|hg|history|host|hostname|htop|iconv|id|ifconfig|ifdown|ifup|import|install|ip|java|jobs|join|kill|killall|less|link|ln|locate|logname|logrotate|look|lpc|lpr|lprint|lprintd|lprintq|lprm|ls|lsof|lynx|make|man|mc|mdadm|mkconfig|mkdir|mke2fs|mkfifo|mkfs|mkisofs|mknod|mkswap|mmv|more|most|mount|mtools|mtr|mutt|mv|nano|nc|netstat|nice|nl|node|nohup|notify-send|npm|nslookup|op|open|parted|passwd|paste|pathchk|ping|pkill|pnpm|podman|podman-compose|popd|pr|printcap|printenv|ps|pushd|pv|quota|quotacheck|quotactl|ram|rar|rcp|reboot|remsync|rename|renice|rev|rm|rmdir|rpm|rsync|scp|screen|sdiff|sed|sendmail|seq|service|sftp|sh|shellcheck|shuf|shutdown|sleep|slocate|sort|split|ssh|stat|strace|su|sudo|sum|suspend|swapon|sync|sysctl|tac|tail|tar|tee|time|timeout|top|touch|tr|traceroute|tsort|tty|umount|uname|unexpand|uniq|units|unrar|unshar|unzip|update-grub|uptime|useradd|userdel|usermod|users|uudecode|uuencode|v|vcpkg|vdir|vi|vim|virsh|vmstat|wait|watch|wc|wget|whereis|which|who|whoami|write|xargs|xdg-open|yarn|yes|zenity|zip|zsh|zypper)(?=$|[)\s;|&])/,lookbehind:!0},keyword:{pattern:/(^|[\s;|&]|[<>]\()(?:case|do|done|elif|else|esac|fi|for|function|if|in|select|then|until|while)(?=$|[)\s;|&])/,lookbehind:!0},builtin:{pattern:/(^|[\s;|&]|[<>]\()(?:\.|:|alias|bind|break|builtin|caller|cd|command|continue|declare|echo|enable|eval|exec|exit|export|getopts|hash|help|let|local|logout|mapfile|printf|pwd|read|readarray|readonly|return|set|shift|shopt|source|test|times|trap|type|typeset|ulimit|umask|unalias|unset)(?=$|[)\s;|&])/,lookbehind:!0,alias:"class-name"},boolean:{pattern:/(^|[\s;|&]|[<>]\()(?:false|true)(?=$|[)\s;|&])/,lookbehind:!0},"file-descriptor":{pattern:/\B&\d\b/,alias:"important"},operator:{pattern:/\d?<>|>\||\+=|=[=~]?|!=?|<<[<-]?|[&\d]?>>|\d[<>]&?|[<>][&=]?|&[>&]?|\|[&|]?/,inside:{"file-descriptor":{pattern:/^\d/,alias:"important"}}},punctuation:/\$?\(\(?|\)\)?|\.\.|[{}[\];\\]/,number:{pattern:/(^|\s)(?:[1-9]\d*|0)(?:[.,]\d+)?\b/,lookbehind:!0}},a.inside=e.languages.bash;for(var r=["comment","function-name","for-or-select","assign-left","parameter","string","environment","function","keyword","builtin","boolean","file-descriptor","operator","punctuation","number"],d=n.variable[1].inside,p=0;p<r.length;p++)d[r[p]]=e.languages.bash[r[p]];e.languages.sh=e.languages.bash,e.languages.shell=e.languages.bash})(g);var we=g;class Ue extends y{static observedAttributes=["count"];static states={count:je};connectedCallback(){this.set("parity",()=>this.get("count")%2?"odd":"even"),this.first(".increment").on("click",()=>this.set("count",(e)=>++e)),this.first(".decrement").on("click",()=>this.set("count",(e)=>--e)),this.first(".count").sync($("count")),this.first(".parity").sync($("parity"))}}Ue.define("my-counter");class ze extends y{connectedCallback(){this.first("span").sync($("name")),this.first("input").on("input",(e)=>this.set("name",e.target.value||void 0))}}ze.define("hello-world");class Ze extends y{connectedCallback(){this.set("active",0);let e=this.querySelectorAll(".slide").length,t=(a,n)=>(a+n+e)%e;this.first(".prev").on("click",()=>this.set("active",(a)=>t(a,-1))),this.first(".next").on("click",()=>this.set("active",(a)=>t(a,1))),this.all(".slide").sync((a,n,r)=>M("active",()=>r===this.get("active"))(a,n)),this.all(".dots span").sync((a,n,r)=>M("active",()=>r===this.get("active"))(a,n))}}Ze.define("my-slider");class qe extends y{static observedAttributes=["collapsed"];static states={collapsed:I};connectedCallback(){let e=this.getAttribute("language")||"html",t=this.querySelector("code");if(t)this.set("code",t.textContent?.trim(),!1),B(()=>{let a=document.createElement("code");a.innerHTML=we.highlight(this.get("code"),we.languages[e],e),L(()=>()=>{Array.from(a.childNodes).filter((n)=>n.nodeType!==Node.COMMENT_NODE).forEach((n)=>n.remove()),Array.from(a.childNodes).forEach((n)=>a.appendChild(n))},[a,"h"])}),this.first(".copy").on("click",async(a)=>{let n=a.target,r=n.textContent,d="success";try{await navigator.clipboard.writeText(t.textContent??"")}catch(p){console.error("Error when trying to use navigator.clipboard.writeText()",p),d="error"}n.set("disabled",!0),n.set("label",this.getAttribute(`copy-${d}`)),setTimeout(()=>{n.set("disabled",!1),n.set("label",r)},d==="success"?1000:3000)}),this.first(".overlay").on("click",()=>this.set("collapsed",!1)),this.self.sync(U("collapsed"))}}qe.define("code-block");class Je extends y{static observedAttributes=["accordion"];static states={accordion:I};static consumedContexts=["media-viewport"];connectedCallback(){super.connectedCallback(),this.set("active",0,!1),setTimeout(()=>{if(this.get("media-viewport"))this.set("accordion",()=>["xs","sm"].includes(this.get("media-viewport")))},0),this.self.sync(U("accordion")),this.first("menu").sync(F("ariaHidden","accordion")),this.all("menu button").on("click",(e,t)=>()=>this.set("active",t)).sync((e,t,a)=>F("ariaPressed",()=>this.get("active")===a)(e,t)),this.all("accordion-panel").pass((e,t)=>({open:()=>this.get("active")===t,collapsible:"accordion"}))}}Je.define("tab-list");class Ye extends y{connectedCallback(){this.set("open",this.hasAttribute("open"),!1),this.set("collapsible",this.hasAttribute("collapsible"),!1),this.self.sync(U("open"),U("collapsible"),F("ariaHidden",()=>!this.get("open")&&!this.get("collapsible"))),this.first("details").sync(F("open"),F("ariaDisabled",()=>!this.get("collapsible")))}}Ye.define("accordion-panel");class Xe extends y{static observedAttributes=["disabled"];static states={disabled:I};connectedCallback(){this.first("button").sync($("label"),F("disabled"))}}Xe.define("input-button");var Ae=(e)=>typeof e==="number",oe=(e,t=!1)=>t?parseInt(e,10):parseFloat(e);class Be extends y{static observedAttributes=["value","description"];static states={value:(e)=>this.isNumber?e.map((t)=>oe(t,this.isInteger)).filter(Number.isFinite):e};connectedCallback(){this.input=this.querySelector("input"),this.isNumber=this.input&&this.input.type==="number",this.isInteger=this.hasAttribute("integer"),this.set("value",this.isNumber?this.input.valueAsNumber:this.input.value,!1),this.set("length",this.input.value.length),this.set("empty",()=>!this.get("length")),this.#t(),this.#a(),this.#n(),this.#r(),this.input.onchange=()=>this.#e(this.isNumber?this.input.valueAsNumber:this.input.value),this.input.oninput=()=>this.set("length",this.input.value.length),B(async()=>{let e=this.get("value"),t=this.getAttribute("validate");if(e&&t)await fetch(`${t}?name=${this.input.name}value=${this.input.value}`).then(async(a)=>{let n=await a.text();this.input.setCustomValidity(n),this.set("error",n)}).catch((a)=>this.set("error",a.message));if(this.isNumber&&!Ae(e))return this.set("value",oe(e,this.isInteger));if(this.isNumber&&!Number.isNaN(e))this.input.value=e})}clear(){this.input.value="",this.set("value",""),this.set("length",0),this.input.focus()}#e=(e)=>{if(this.set("value",e),this.set("error",this.input.validationMessage),typeof e==="function")e=this.get("value");if(this.input.value!==String(e))this.dispatchEvent(new CustomEvent("value-change",{detail:e,bubbles:!0}))};#t(){let e=this.first(".error");this.set("ariaInvalid",()=>String(Boolean(this.get("error")))),this.set("aria-errormessage",()=>this.get("error")?e[0]?.target.id:void 0),e.forEach($("error")),this.first("input").map(F("ariaInvalid")).forEach(W("aria-errormessage"))}#a(){let e=this.first(".description");if(!e[0])return;let t=this.first("input"),a=this.input.maxLength,n=a&&e[0].target.dataset.remaining,r=e[0].target.textContent;this.set("description",n?()=>{let d=this.get("length");return d>0?n.replace("${x}",a-d):r}:r),this.set("aria-describedby",()=>this.get("description")?e[0].target.id:void 0),e.forEach($("description")),t.forEach(W("aria-describedby"))}#n(){let e=this.querySelector(".spinbutton");if(!this.isNumber||!e)return;let t=(f)=>maybe(oe(this.input[f],this.isInteger)).filter(Number.isFinite)[0],a=oe(e.dataset.step,this.isInteger),[n,r,d]=Number.isFinite(a)?[a,t("min"),t("max")]:[],p=(f)=>{let b=Math.round((d-r)/n),A=Math.round((f-r)*b/(d-r))/b;A=Math.min(Math.max(A,0),1);let S=A*(d-r)+r;return this.isInteger?Math.round(S):S};this.stepDown=(f=n)=>this.#e((b)=>p(b-f)),this.stepUp=(f=n)=>this.#e((b)=>p(b+f)),this.set("decrement-disabled",()=>Ae(r)&&this.get("value")-n<r),this.set("increment-disabled",()=>Ae(d)&&this.get("value")+n>d),this.first(".decrement").map(F("disabled","decrement-disabled")).forEach(on("click",(f)=>this.stepDown(f.shiftKey?n*10:n))),this.first(".increment").map(F("disabled","increment-disabled")).forEach(on("click",(f)=>this.stepUp(f.shiftKey?n*10:n))),this.input.onkeydown=(f)=>{if(["ArrowUp","ArrowDown"].includes(f.key)){if(f.stopPropagation(),f.preventDefault(),f.key==="ArrowDown")this.stepDown(f.shiftKey?n*10:n);if(f.key==="ArrowUp")this.stepUp(f.shiftKey?n*10:n)}}}#r(){this.first(".clear").map(M("hidden","empty")).forEach(on("click",()=>{this.clear(),this.input.focus()}))}}Be.define("input-field");class Ie extends y{static observedAttributes=["checked"];static states={checked:I};connectedCallback(){this.first("input").on("change",(e)=>this.set("checked",Boolean(e.target.checked))).sync(F("checked")),this.self.sync(U("checked"))}}Ie.define("input-checkbox");class Qe extends y{static observedAttributes=["value"];connectedCallback(){this.self.sync(W("value")),this.all("input").on("change",(e)=>this.set("value",e.target.value)),this.all("label").sync((e,t)=>M("selected",()=>e.get("value")===t.querySelector("input")?.value)(e,t))}}Qe.define("input-radiogroup");class We extends y{static observedAttributes=["src"];static states={src:(e)=>e.map((t)=>{let a="";try{if(a=new URL(t,location.href),a.origin!==location.origin)throw new TypeError("Invalid URL origin")}catch(n){console.error(n,a),a=""}return a.toString()})};connectedCallback(){this.first(".loading").forEach(F("ariaHidden",()=>!!this.get("error"))),this.first(".error").map($("error")).forEach(F("ariaHidden",()=>!this.get("error"))),B((e)=>{let t=this.get("src");if(!t)return;fetch(t).then(async(a)=>{if(a.ok){let n=await a.text();e(this.root,"h",(r)=>()=>{r.innerHTML=n,r.querySelectorAll("script").forEach((d)=>{let p=document.createElement("script");p.appendChild(document.createTextNode(d.textContent)),r.appendChild(p),d.remove()})}),this.set("error","")}else this.set("error",a.status+":"+a.statusText)}).catch((a)=>this.set("error",a))})}}We.define("lazy-load");var xe="media-motion",ue="media-theme",z="media-viewport",He="media-orientation";class Me extends y{static providedContexts=[xe,ue,z,He];connectedCallback(){super.connectedCallback();let e="light",t="dark",a="xs",n="sm",r="md",d="lg",p="xl",f="landscape",b="portrait",S=(()=>{let c=(j)=>{let T=this.getAttribute(j)?.trim();if(!T)return null;let K=T.match(/em$/)?"em":"px",V=maybe(parseFloat(T)).filter(Number.isFinite)[0];return V?V+K:null},h=c(n)||"32em",m=c(r)||"48em",k=c(d)||"72em",v=c(p)||"108em";return{sm:h,md:m,lg:k,xl:v}})(),N=matchMedia("(prefers-reduced-motion: reduce)"),G=matchMedia("(prefers-color-scheme: dark)"),Z=matchMedia(`(min-width: ${S.sm})`),i=matchMedia(`(min-width: ${S.md})`),s=matchMedia(`(min-width: ${S.lg})`),u=matchMedia(`(min-width: ${S.xl})`),l=matchMedia("(orientation: landscape)"),o=()=>{if(u.matches)return p;if(s.matches)return d;if(i.matches)return r;if(Z.matches)return n;return a};this.set(xe,N.matches),this.set(ue,G.matches?t:e),this.set(z,o()),this.set(He,l.matches?f:b),N.addEventListener("change",(c)=>this.set(xe,c.matches)),G.addEventListener("change",(c)=>this.set(ue,c.matches?t:e)),Z.addEventListener("change",()=>this.set(z,o())),i.addEventListener("change",()=>this.set(z,o())),s.addEventListener("change",()=>this.set(z,o())),u.addEventListener("change",()=>this.set(z,o())),l.addEventListener("change",(c)=>this.set(ue,c.matches?f:b))}}Me.define("media-context");class Ke extends y{static states={tasks:[],filter:"all"};connectedCallback(){this.#e(),this.self.on("click",(e)=>{if(e.target.localName==="button")this.removeItem(e.target)}).sync(W("filter")),this.set("count",()=>{let e=this.get("tasks").map((n)=>n.signals.get("checked")),t=e.filter((n)=>n?.get()).length,a=e.length;return{active:a-t,completed:t,total:a}})}addItem=(e)=>{let t=this.querySelector("template")?.content.cloneNode(!0);if(t&&t.querySelector("span"))t.querySelector("span").textContent=e,this.querySelector("ol")?.appendChild(t),this.#e()};removeItem=(e)=>{e.closest("li")?.remove(),this.#e()};clearCompleted=()=>{this.get("tasks").filter((e)=>e.get("checked")).forEach((e)=>e.parentElement?.remove()),this.#e()};#e(){this.set("tasks",Array.from(this.querySelectorAll("input-checkbox")))}}Ke.define("todo-list");class Ve extends y{connectedCallback(){let e=this.querySelector("todo-list"),t=this.querySelector("input-radiogroup");this.self.on("add-todo",(a)=>e?.addItem(a.detail)),this.first("todo-count").pass({active:()=>(e?.get("count")).active}),this.first("todo-list").pass({filter:()=>t?.get("value")}),this.first(".clear-completed").on("click",()=>e?.clearCompleted()).pass({disabled:()=>!(e?.get("count")).completed})}}Ve.define("todo-app");class et extends y{static states={active:0};connectedCallback(){this.first(".count").sync($("active")),this.first(".singular").sync(F("ariaHidden",()=>this.get("active")>1)),this.first(".plural").sync(F("ariaHidden",()=>this.get("active")===1)),this.first(".remaining").sync(F("ariaHidden",()=>!this.get("active"))),this.first(".all-done").sync(F("ariaHidden",()=>!!this.get("active")))}}et.define("todo-count");class tt extends y{connectedCallback(){let e=this.querySelector("input-field");this.first("form").on("submit",(t)=>{t.preventDefault(),setTimeout(()=>{this.dispatchEvent(new CustomEvent("add-todo",{bubbles:!0,detail:e?.get("value")})),e?.clear()},0)}),this.first("input-button").pass({disabled:()=>e?.get("empty")})}}tt.define("todo-form");export{Qe as InputRadiogroup};
