+++ docs/changelog-versioning.html
@@ -5,11 +5,11 @@
 		<link rel="stylesheet" href="assets/css/global.css">
 		<link rel="stylesheet" href="assets/css/okaidia.css">
 		<link rel="stylesheet" href="assets/css/components.css">
-		<script type="module" src="assets/js/main.min.js"></script>
+		<script type="module" src="assets/js/main.js"></script>
 	</head>
 	<body>
 		<header class="content-grid">
-			<h1 class="content">UIElement Docs <small>Version 0.8.5</small></h1>
+			<h1 class="content">UIElement Docs <small>Version 0.9.4</small></h1>
 			<nav class="breakout">
 				<ol>
 					<li>
@@ -174,7 +174,7 @@
 					</li>
 					<li>Breaking Changes to Core Features since v0.7.3:
 						<ul>
-						<li><code>attributeMap</code> is now static like <code>observedAttributes</code>.</li>
+						<li><code>states</code> is now static like <code>observedAttributes</code>.</li>
 						<li><code>this.pass(element, stateMap)</code> is replaced by <code>this.[first|all](selector).map(pass(stateMap))</code>.</li>
 						</ul>
 					</li>
@@ -191,7 +191,7 @@
 					</li>
 					<li>Breaking Changes since 0.7.2:
 						<ul>
-						<li>Removed support for array values in <code>attributeMap</code>.</li>
+						<li>Removed support for array values in <code>states</code>.</li>
 						<li>Refactored component props interface.</li>
 						</ul>
 					</li>
@@ -254,7 +254,7 @@
 					</li>
 					<li>Breaking Changes since v0.4.0:
 						<ul>
-						<li><code>attributeMapping</code> renamed to <code>attributeMap</code>.</li>
+						<li><code>statesping</code> renamed to <code>states</code>.</li>
 						</ul>
 					</li>
 					</ul>
