+++ docs/core-concepts.html
@@ -5,11 +5,11 @@
 		<link rel="stylesheet" href="assets/css/global.css">
 		<link rel="stylesheet" href="assets/css/okaidia.css">
 		<link rel="stylesheet" href="assets/css/components.css">
-		<script type="module" src="assets/js/main.min.js"></script>
+		<script type="module" src="assets/js/main.js"></script>
 	</head>
 	<body>
 		<header class="content-grid">
-			<h1 class="content">UIElement Docs <small>Version 0.8.5</small></h1>
+			<h1 class="content">UIElement Docs <small>Version 0.9.4</small></h1>
 			<nav class="breakout">
 				<ol>
 					<li>
@@ -200,12 +200,12 @@ this.set('isEven', () => this.get('count') % 2 === 0); // Create a derived signa
 						<button type="button" class="secondary small">Copy</button>
 					</input-button>
                 </code-block>
-				<p>To parse and map attributes to specific state types, use the <code>attributeMap</code> static property.</p>
+				<p>To parse and map attributes to specific state types, use the <code>states</code> static property.</p>
 				<code-block language="js" copy-success="Copied!" copy-error="Error trying to copy to clipboard!">
 					<p class="meta">
 						<span class="language">js</span>
 					</p>
-					<pre class="language-js"><code>static attributeMap = {
+					<pre class="language-js"><code>static states = {
 	count: asInteger,
 }</code></pre>
 					<input-button class="copy">
@@ -217,7 +217,7 @@ this.set('isEven', () => this.get('count') % 2 === 0); // Create a derived signa
 					<p class="meta">
 						<span class="language">js</span>
 					</p>
-				<pre class="language-js"><code>static attributeMap = {
+				<pre class="language-js"><code>static states = {
 	date: value => value.map(v => new Date(v)),
 }</code></pre>
 					<input-button class="copy">
