+++ docs/detailed-walkthrough.html
@@ -5,11 +5,11 @@
 		<link rel="stylesheet" href="assets/css/global.css">
 		<link rel="stylesheet" href="assets/css/okaidia.css">
 		<link rel="stylesheet" href="assets/css/components.css">
-		<script type="module" src="assets/js/main.min.js"></script>
+		<script type="module" src="assets/js/main.js"></script>
 	</head>
 	<body>
 		<header class="content-grid">
-			<h1 class="content">UIElement Docs <small>Version 0.8.5</small></h1>
+			<h1 class="content">UIElement Docs <small>Version 0.9.4</small></h1>
 			<nav class="breakout">
 				<ol>
 					<li>
@@ -111,7 +111,7 @@
 				<p>State in a component can come from several sources, each with its own way of being initialized and managed.</p>
 			
 				<h3>Observed Attributes</h3>
-				<p>Attributes declared in <code>static observedAttributes</code> are automatically parsed and converted into reactive signals using <code>static attributeMap</code>.</p>
+				<p>Attributes declared in <code>static observedAttributes</code> are automatically parsed and converted into reactive signals using <code>static states</code>.</p>
 				<code-block language="js" copy-success="Copied!" copy-error="Error trying to copy to clipboard!">
 					<p class="meta">
 						<span class="language">js</span>
@@ -120,7 +120,7 @@
 
 class CounterComponent extends UIElement {
 	static observedAttributes = ['count']
-	static attributeMap = { count: asInteger }
+	static states = { count: asInteger }
 }</code></pre>
 					<input-button class="copy">
 						<button type="button" class="secondary small">Copy</button>
@@ -141,7 +141,7 @@ class CounterComponent extends UIElement {
 					<li><code>asBoolean</code>: Convert boolean attributes like <code>disabled</code> to actual booleans.</li>
 					<li><code>asInteger</code>: Try to parse numeric attributes as integer numbers.</li>
 					<li><code>asNumber</code>: Try to parse numeric attributes as floating point numbers.</li>
-					<li><code>asString</code>: Does nothing as all attributes are strings; string attributes may be omitted in <code>attributeMap</code>.</li>
+					<li><code>asString</code>: Does nothing as all attributes are strings; string attributes may be omitted in <code>states</code>.</li>
 					<li><code>asJSON</code>: Try to parse structured data in an attribute as JSON.</li>
 				</ul>
 				<p>If the parser fails, for example because numeric attribute evaluate to <code>NaN</code> or the supplied JSON is invalid, the signal will silently default to <code>undefined</code>, just as if the attribute wasn't there. Undefined signals won't trigger any effects.</p>
