+++ docs/examples/code-block.html
@@ -159,7 +159,7 @@ import Prism from 'prismjs'
 
 class CodeBlock extends UIElement {
 	static observedAttributes = ['collapsed']
-	static attributeMap = {
+	static states = {
 		collapsed: asBoolean
 	}
