+++ docs/examples/input-button.html
@@ -159,7 +159,7 @@
 
 class InputButton extends UIElement {
 	static observedAttributes = ['disabled']
-	static attributeMap = {
+	static states = {
 		disabled: asBoolean
 	}
