+++ docs/examples/input-checkbox.html
@@ -123,7 +123,7 @@
 
 class InputCheckbox extends UIElement {
 	static observedAttributes = ['checked']
-	static attributeMap = {
+	static states = {
 		checked: asBoolean,
 	}
