+++ docs/examples/my-counter.html
@@ -78,7 +78,7 @@
 
 class MyCounter extends UIElement {
 	static observedAttributes = ['count']
-	static attributeMap = {
+	static states = {
 		count: asInteger
 	}
