+++ docs/examples/tab-list.html
@@ -93,7 +93,7 @@
 
 class TabList extends UIElement {
 	static observedAttributes = ['accordion']
-	static attributeMap = {
+	static states = {
 		accordion: asBoolean
 	}
 	static consumedContexts = ['media-viewport']
