+++ src/core/parse.ts
@@ -13,7 +13,7 @@ const parseNumber = (parseFn: (v: string) => number, value?: string): number | u
 /* === Exported Functions === */
 
 /**
- * Parse according to static attributeMap
+ * Parse according to static states
  * 
  * @since 0.8.4
  * @param {UIElement} host - host UIElement
